requirements.txt
keepa==1.3.13
pandas==2.2.2
numpy==2.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
python-dotenv==1.0.1
scikit-learn==1.5.1
fastapi==0.112.2
uvicorn==0.30.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
